1. Sorting keys from a map[string]string to insert them in a
map[string][]string makes little sense.
2. https://golang.org/pkg/net/url/#Values.Encode does sort the keys
3. On the server side we no longer sort the keys, see https://github.com/thoas/picfit/commit/3530bf29c1ef1dbef96d5179055f8c776ceffe1a
4. Here is the speedup:
```bash
    # before
    $ go test -bench=.
    goos: linux
    goarch: amd64
    pkg: github.com/ulule/picfit-go
    BenchmarkBuildParams-4   	  200000	      5735 ns/op
    PASS
    ok  	github.com/ulule/picfit-go	1.238s
    # after
    $ go test -bench=.
    goos: linux
    goarch: amd64
    pkg: github.com/ulule/picfit-go
    BenchmarkBuildParams-4   	 3000000	       548 ns/op
    PASS
    ok  	github.com/ulule/picfit-go	2.212s
```